avatar_refactoring/requirements.txt
asgiref==3.9.1
boto3==1.39.3
botocore==1.39.3
Django==5.2.4
django-rest-framework==0.1.0
django-storages==1.14.6
djangorestframework==3.16.0
jmespath==1.0.1
pillow==11.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
